+++ CHANGELOG.md
@@ -10,6 +10,8 @@
   - Payloads:
     - Remove redundant/unused files bundled in payloads.dmg
 - Resolve unpatching Nvidia Web Drivers failing to clean up `/Library/Extensions`
+- Implement preflight code signature checks for macOS installer creation
+  - Ensures validity of `createinstallmedia` binary before execution
 
 ## 1.4.3
 - Update non-Metal Binaries for macOS Sonoma: